Break out PartDesign tests into own folder
