+++ src/Mod/PartDesign/PartDesignTests/TestPad.py
@@ -1,0 +1,158 @@
+#   (c) Juergen Riegel (FreeCAD@juergen-riegel.net) 2011      LGPL        *
+#                                                                         *
+#   This file is part of the FreeCAD CAx development system.              *
+#                                                                         *
+#   This program is free software; you can redistribute it and/or modify  *
+#   it under the terms of the GNU Lesser General Public License (LGPL)    *
+#   as published by the Free Software Foundation; either version 2 of     *
+#   the License, or (at your option) any later version.                   *
+#   for detail see the LICENCE text file.                                 *
+#                                                                         *
+#   FreeCAD is distributed in the hope that it will be useful,            *
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+#   GNU Library General Public License for more details.                  *
+#                                                                         *
+#   You should have received a copy of the GNU Library General Public     *
+#   License along with FreeCAD; if not, write to the Free Software        *
+#   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  *
+#   USA                                                                   *
+#**************************************************************************
+import unittest
+
+import FreeCAD
+import TestSketcherApp
+
+class TestPad(unittest.TestCase):
+    def setUp(self):
+        self.Doc = FreeCAD.newDocument("PartDesignTestPad")
+
+    def testBoxCase(self):
+        self.PadSketch = self.Doc.addObject('Sketcher::SketchObject','SketchPad')
+        TestSketcherApp.CreateSlotPlateSet(self.PadSketch)
+        self.Doc.recompute()
+        self.Pad = self.Doc.addObject("PartDesign::Pad","Pad")
+        self.Pad.Profile = self.PadSketch
+        self.Doc.recompute()
+        self.assertEqual(len(self.Pad.Shape.Faces), 6)
+
+    def testPadToFirstCase(self):
+        self.Body = self.Doc.addObject('PartDesign::Body','Body')
+        # Make first offset cube Pad
+        self.PadSketch = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad')
+        self.Body.addObject(self.PadSketch)
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch, (0, 1), (1, 1))
+        self.Doc.recompute()
+        self.Pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+        self.Body.addObject(self.Pad)
+        self.Pad.Profile = self.PadSketch
+        self.Pad.Length = 1
+        self.Doc.recompute()
+        # Make second pad on different plane and pad to first
+        self.PadSketch1 = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad1')
+        self.Body.addObject(self.PadSketch1)
+        self.PadSketch1.MapMode = 'FlatFace'
+        self.PadSketch1.Support = (self.Doc.XZ_Plane, [''])
+        self.Doc.recompute()
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch1, (0, 0), (1, 1))
+        self.Doc.recompute()
+        self.Pad1 = self.Doc.addObject("PartDesign::Pad", "Pad1")
+        self.Body.addObject(self.Pad1)
+        self.Pad1.Profile = self.PadSketch1
+        self.Pad1.Type = 2
+        self.Pad1.Reversed = 1
+        self.Doc.recompute()
+        self.assertAlmostEqual(self.Pad1.Shape.Volume, 2.0)
+
+    def testPadtoLastCase(self):
+        self.Body = self.Doc.addObject('PartDesign::Body','Body')
+        # Make first offset cube Pad
+        self.PadSketch = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad')
+        self.Body.addObject(self.PadSketch)
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch, (0.5, 1), (0.5, 2))
+        self.Doc.recompute()
+        self.Pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+        self.Body.addObject(self.Pad)
+        self.Pad.Profile = self.PadSketch
+        self.Pad.Length = 1
+        self.Doc.recompute()
+        # Make second pad on different plane and pad to first
+        self.PadSketch1 = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad1')
+        self.Body.addObject(self.PadSketch1)
+        self.PadSketch1.MapMode = 'FlatFace'
+        self.PadSketch1.Support = (self.Doc.XZ_Plane, [''])
+        self.Doc.recompute()
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch1, (0, 0), (1, 1))
+        self.Doc.recompute()
+        self.Pad1 = self.Doc.addObject("PartDesign::Pad", "Pad1")
+        self.Body.addObject(self.Pad1)
+        self.Pad1.Profile = self.PadSketch1
+        self.Pad1.Type = 1
+        self.Pad1.Reversed = 1
+        self.Doc.recompute()
+        self.assertAlmostEqual(self.Pad1.Shape.Volume, 3.0)
+
+    def testPadToFaceCase(self):
+        self.Body = self.Doc.addObject('PartDesign::Body','Body')
+        # Make first offset cube Pad
+        self.PadSketch = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad')
+        self.Body.addObject(self.PadSketch)
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch, (0, 1), (1, 1))
+        self.Doc.recompute()
+        self.Pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+        self.Body.addObject(self.Pad)
+        self.Pad.Profile = self.PadSketch
+        self.Pad.Length = 1
+        self.Doc.recompute()
+        # Make second pad on different plane and pad to face on first
+        self.PadSketch1 = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad1')
+        self.Body.addObject(self.PadSketch1)
+        self.PadSketch1.MapMode = 'FlatFace'
+        self.PadSketch1.Support = (self.Doc.XZ_Plane, [''])
+        self.Doc.recompute()
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch1, (0, 0), (1, 1))
+        self.Doc.recompute()
+        self.Pad1 = self.Doc.addObject("PartDesign::Pad", "Pad1")
+        self.Body.addObject(self.Pad1)
+        self.Pad1.Profile = self.PadSketch1
+        self.Pad1.Type = 3
+        self.Pad1.UpToFace = (self.Pad, ["Face3"])
+        self.Pad1.Reversed = 1
+        self.Doc.recompute()
+        self.assertAlmostEqual(self.Pad1.Shape.Volume, 2.0)
+
+    def testPadTwoDimensionsCase(self):
+        self.Body = self.Doc.addObject('PartDesign::Body','Body')
+        # Make first offset cube Pad
+        self.PadSketch = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad')
+        self.Body.addObject(self.PadSketch)
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch, (0, 1), (1, 1))
+        self.Doc.recompute()
+        self.Pad = self.Doc.addObject("PartDesign::Pad", "Pad")
+        self.Body.addObject(self.Pad)
+        self.Pad.Profile = self.PadSketch
+        self.Pad.Length = 1
+        self.Doc.recompute()
+        # Make second pad on different plane and pad to face on first
+        self.PadSketch1 = self.Doc.addObject('Sketcher::SketchObject', 'SketchPad1')
+        self.Body.addObject(self.PadSketch1)
+        self.PadSketch1.MapMode = 'FlatFace'
+        self.PadSketch1.Support = (self.Doc.XZ_Plane, [''])
+        self.Doc.recompute()
+        TestSketcherApp.CreateRectangleSketch(self.PadSketch1, (0, 0), (1, 1))
+        self.Doc.recompute()
+        self.Pad1 = self.Doc.addObject("PartDesign::Pad", "Pad1")
+        self.Body.addObject(self.Pad1)
+        self.Pad1.Profile = self.PadSketch1
+        self.Pad1.Type = 4
+        self.Pad1.Length = 2.0
+        self.Pad1.Length2 = 1.0
+        self.Pad1.Reversed = 1
+        self.Doc.recompute()
+        self.assertAlmostEqual(self.Pad1.Shape.Volume, 4.0)
+
+    def tearDown(self):
+        #closing doc
+        FreeCAD.closeDocument("PartDesignTestPad")
+        #print ("omit closing document for debugging")
+
